Move Embedd classes to subconstruct.py
